Made Many Changes (See full commit messagee)
Added a new table underneath the page content to correspond to Badge
Eligibility in AYSO. Also took the repetitive JS and moved it to the
`main.js` file, in order to make this website more efficient and save
space. Tried to make the editGame.html page scrollable...hopefully
it works. Also in a previous commit, added a couple of lines to the
`sortDate()` function in order to move all of the empty date to the
bottom when sorted. And in another previous commit, changed database
methods from `.update()` to `.set()` so that it overwrites the old
`gamesList` object with the new `gamesList` object, thus deleting
and game that needs to be deleted.